+++ src/Automated/Assertion/Property/CountsEqual.php
@@ -47,7 +47,7 @@ class CountsEqual extends AbstractConstraint
 {
     use ImmutableTrait;
 
-    private int $exceptedCount;
+    private readonly int $exceptedCount;
 
     public function __construct(int $exceptedCount)
     {