+++ src/Automated/Assertion/Property/HasEmptyValueForKey.php
@@ -45,7 +45,7 @@ class HasEmptyValueForKey extends AbstractConstraint
 {
     use ImmutableTrait;
 
-    private string $keyName;
+    private readonly string $keyName;
 
     public function __construct(string $keyName)
     {