+++ src/State/StateInterface.php
@@ -57,21 +57,6 @@ use Teknoo\States\Proxy\ProxyInterface;
 interface StateInterface
 {
     /**
-     * Const to get a closure into a public scope.
-     */
-    public const VISIBILITY_PUBLIC = 'u';
-
-    /**
-     * Const to get a closure into a protected scope.
-     */
-    public const VISIBILITY_PROTECTED = 'r';
-
-    /**
-     * Const to get a closure into a private scope.
-     */
-    public const VISIBILITY_PRIVATE = 'i';
-
-    /**
      * To initialize this state.
      *
      * @api
@@ -106,7 +91,7 @@ interface StateInterface
      * @param ProxyInterface $object the instance to use to bind with an object's scope the closure
      * @param string         $methodName
      * @param array<mixed>   $arguments to pass to the closure
-     * @param string         $requiredScope self::VISIBILITY_PUBLIC|self::VISIBILITY_PROTECTED|self::VISIBILITY_PRIVATE
+     * @param Visibility     $requiredScope
      * @param string         $statedClassOrigin
      * @param callable       $returnCallback Method to call if the closure has been found and called, to pass the result
      */
@@ -114,7 +99,7 @@ interface StateInterface
         ProxyInterface $object,
         string &$methodName,
         array &$arguments,
-        string &$requiredScope,
+        Visibility &$requiredScope,
         string &$statedClassOrigin,
         callable &$returnCallback
     ): StateInterface;